XML format update
- Changed around the HazeronWatcherSettings values a bit and made some
nested classes. This is all just so that the settings XML file is easier
to read through manually.
- Renamed "Settings.xml" to "HazeronWatcherSettings.xml".
- Added in some code to convert a "Settings.xml" file into a
"HazeronWatcherSettings.xml" file on startup. The old "Settings.xml"
file is renamed to "Settings (v1.1 backup).xml".
- Default values are now set in the
HazeronWatcher.HazeronWatcherSettingsOptions constructor. It isn't a
pretty place to do it, but it works for now.