MainID WIP
This feature is still not finished, having a number of problems that I
am unsure how to really fix right now.
But I am uploading the code I have right now.

Note this is only the code to add the MainID to a character, there is no
actual function for what to do with this MainID yet.

Also:
- Moved the MainID XmlAttribute to under Name